reduced number of matrix calculations per frame

+++ SRC/TRIANGLE.C
@@ -63,12 +63,12 @@ void gfx_drawTriangle(const gfx_Triangle *t, const mth_Matrix4 *matrix, gfx_draw
     }
 
     // transform x and y of each vertex to screen coordinates
-    v0.position.x = (v0.position.x * (float)buffer->width)  / (2.0f * v0.position.w) + bufferHalfWidth;
-    v0.position.y = (v0.position.y * (float)buffer->height) / (2.0f * v0.position.w) + bufferHalfHeight;
-    v1.position.x = (v1.position.x * (float)buffer->width)  / (2.0f * v1.position.w) + bufferHalfWidth;
-    v1.position.y = (v1.position.y * (float)buffer->height) / (2.0f * v1.position.w) + bufferHalfHeight;
-    v2.position.x = (v2.position.x * (float)buffer->width)  / (2.0f * v2.position.w) + bufferHalfWidth;
-    v2.position.y = (v2.position.y * (float)buffer->height) / (2.0f * v2.position.w) + bufferHalfHeight;
+    v0.position.x = (v0.position.x * buffer->width)  / (2.0 * v0.position.w) + bufferHalfWidth;
+    v0.position.y = (v0.position.y * buffer->height) / (2.0 * v0.position.w) + bufferHalfHeight;
+    v1.position.x = (v1.position.x * buffer->width)  / (2.0 * v1.position.w) + bufferHalfWidth;
+    v1.position.y = (v1.position.y * buffer->height) / (2.0 * v1.position.w) + bufferHalfHeight;
+    v2.position.x = (v2.position.x * buffer->width)  / (2.0 * v2.position.w) + bufferHalfWidth;
+    v2.position.y = (v2.position.y * buffer->height) / (2.0 * v2.position.w) + bufferHalfHeight;
 
     // sort vertices so that v0 is topmost, then v2, then v1
     if(v2.position.y > v1.position.y)
